SCRUM-42 feat: adicionado feature de apagar todas as orientações

+++ src/aplication/interfaces/MenuVisualizarOrientacao.java
@@ -61,21 +61,48 @@ public class MenuVisualizarOrientacao extends Menu {
 	}
 
 	public void removerOrientacao(Scanner input, MenuHistorico menuHistorico) {
-		try {
-			menuHistorico.sobscreverMenu(menuHistorico.pegarMenuAnterior());
-			idiomaImplementacao.mostrarMenuConfirmarApagarOrientacao(input);
-
-			orientacaoService.removerOrientacao(orientacaoDto);
+		try {			
+			escolherOpcaoDeletar(input, menuHistorico);
+			
 			var proximoMenu = MenuFactory.criarMenuResultado(TipoMenu.CERTO, menuHistorico.voltarMenu(),
 					idiomaImplementacao.pegarMensagemRemoverComSucessoOrientacao(), idiomaImplementacao);
 			menuHistorico.definirProximoMenu(proximoMenu);
 		} catch (SairMenuException sme) {
-			menuHistorico.voltarMenu();
+			System.out.println(idiomaImplementacao.pegarMensagemVoltandoMenu());
 		} catch (Exception le) {
 			var proximoMenu = MenuFactory.criarMenuResultado(TipoMenu.FALHA, menuHistorico.voltarMenu(),
 					idiomaImplementacao.pegarMensagemErroAoRemoverOrientacao(), idiomaImplementacao);
 			menuHistorico.definirProximoMenu(proximoMenu);
 		}
+	}
+
+	private void escolherOpcaoDeletar(Scanner input, MenuHistorico menuHistorico) throws SairMenuException {
+		String opcaoDelete = idiomaImplementacao.mostrarMenuOpcaoApagarOrientacao(input);
+		
+		switch(opcaoDelete.trim().toUpperCase()) {
+		case "T":
+			if(confirmarSeDesejaApagar(input)) {
+				menuHistorico.sobscreverMenu(menuHistorico.pegarMenuAnterior());
+				orientacaoService.apagarOrientacaoId(orientacaoDto);
+			}
+			break;
+		case "A":
+			if(confirmarSeDesejaApagar(input)) {
+				menuHistorico.sobscreverMenu(menuHistorico.pegarMenuAnterior());
+				orientacaoService.removerOrientacao(orientacaoDto);
+			}
+			break;
+		}
+		
+	}
+	
+	private boolean confirmarSeDesejaApagar(Scanner input) throws SairMenuException {
+		String confirmar = idiomaImplementacao.mostrarMenuConfirmarApagarOrientacao(input);
+		
+		if(confirmar.trim().toUpperCase().equals("A")) {
+			return true;
+		}
+		return false;
 	}
 
 	public List<IdiomaOrientacao> gerarListaOrdenada(Map<IdiomaOrientacao, OrientacaoDto> listaComOrientacoes) {
